Split configuration deserialization compile unit
